requirements.txt
torch~=1.8.1+cu102
matplotlib~=3.5.1
numpy~=1.21.5
torchvision~=0.9.1+cu102
tqdm~=4.62.3
Pillow~=9.0.1